rag-analytics-agent/requirements.txt
# ─── LLM + Embeddings ─────────────────────────────────────────────────────────
# openai provides two things:
#   1. text-embedding-3-small  → converts text to vectors for ChromaDB
#   2. gpt-4o                  → generates SQL and formats answers
openai>=1.30.0

# ─── Vector Store ─────────────────────────────────────────────────────────────
# ChromaDB stores the schema embeddings locally (no external service needed).
# Persists to ./chroma_db/ on disk so the index survives restarts.
chromadb>=0.5.0

# ─── Mock Data Warehouse ──────────────────────────────────────────────────────
# DuckDB is the local SQL engine used in mock mode (USE_MOCK_DB=true).
# It speaks standard SQL and is fully compatible with the queries GPT-4o generates.
duckdb>=0.10.0

# pandas is used by seed/mock_data.py for fast bulk inserts into DuckDB
pandas>=2.0.0

# ─── Production Data Warehouse (optional) ─────────────────────────────────────
# Uncomment if USE_MOCK_DB=false (connecting to a real Snowflake account)
# snowflake-connector-python>=3.10.0

# ─── API Framework ────────────────────────────────────────────────────────────
# FastAPI: the web framework
# uvicorn: the ASGI server that runs FastAPI (standard = includes websocket support)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ─── Utilities ────────────────────────────────────────────────────────────────
# python-dotenv: reads .env file into os.environ
python-dotenv>=1.0.0
# pydantic: data validation for FastAPI request/response models (bundled with fastapi)
pydantic>=2.0.0
